back-end/Sign/requirements.txt
backports.entry-points-selectable==1.1.0
click==8.0.1
cycler==0.10.0
distlib==0.3.3
filelock==3.2.0
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
nltk==3.6.5
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.4.0
platformdirs==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
regex==2021.10.8
six==1.16.0
tqdm==4.62.3
virtualenv==20.8.1
Werkzeug==2.0.1
